+++ src/components/common/definitions/defineAllComponents.ts
@@ -63,6 +63,7 @@ import IgcTextareaComponent from '../../textarea/textarea.js';
 import IgcTileManagerComponent from '../../tile-manager/tile-manager.js';
 import IgcTileComponent from '../../tile-manager/tile.js';
 import IgcToastComponent from '../../toast/toast.js';
+import IgcTooltipComponent from '../../tooltip/tooltip.js';
 import IgcTreeItemComponent from '../../tree/tree-item.js';
 import IgcTreeComponent from '../../tree/tree.js';
 import { defineComponents } from './defineComponents.js';
@@ -136,6 +137,7 @@ const allComponents: IgniteComponent[] = [
   IgcTextareaComponent,
   IgcTileComponent,
   IgcTileManagerComponent,
+  IgcTooltipComponent,
 ];
 
 export function defineAllComponents() {