+++ src/components/common/utils.spec.ts
@@ -202,6 +202,14 @@ export function simulatePointerLeave(
   );
 }
 
+export function simulateFocus(node: Element) {
+  node.dispatchEvent(new FocusEvent('focus'));
+}
+
+export function simulateBlur(node: Element) {
+  node.dispatchEvent(new FocusEvent('blur'));
+}
+
 export function simulatePointerDown(
   node: Element,
   options?: PointerEventInit,